spawn process/message, warn on large message queue
Under even relatively moderate load, the bridge may become overwhelmed
and develop a backlog in its message queue that it can not clear for
long. This will eventually crash the node due to oom.

This is a quickfix solution to keep the bridge processs light weight
and to offload the transformation to a new process per message.
It is not ideal since all messages go through the bridge process,
but storing them off heap may help.

This patch may be entirely or selectively reverted if it does not
perform as expected.